Use rev-parse to fetch branch and upstream names
This commit improves the performance of git-prompt. Previously
'git status' was used to extract the names of the local and remote
branches. This could potentially take a lot of time, since 'git status'
reviews the current status of the repository and is quite slow if there
are a lot of non staged files. This update makes use of 'git rev-parse'
to directly get the names of the local and remote branches from the
index.